// osa9/courseinformation/src/types.tsx
interface CoursePartBase {
  name: string
  exerciseCount: number
}

interface CoursePartBaseWithDescription extends CoursePartBase {
  description?: string
}

interface CoursePartOne extends CoursePartBaseWithDescription {
  name: "Fundamentals"
}

interface CoursePartTwo extends CoursePartBase {
  name: "Using props to pass data"
  groupProjectCount: number
}

interface CoursePartThree extends CoursePartBaseWithDescription {
  name: "Deeper type usage"
  exerciseSubmissionLink: string
}

interface CoursePartFour extends CoursePartBaseWithDescription {
  name: "This is deep"
  description: string
}

export type CoursePart = CoursePartOne | CoursePartTwo | CoursePartThree | CoursePartFour